% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.breakpoint_boot}
\alias{plot.breakpoint_boot}
\title{Plot the bootstrap distribution of a breakpoint}
\usage{
\method{plot}{breakpoint_boot}(x, ...)
}
\arguments{
\item{x}{A \code{breakpoint_boot} object.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Plot the bootstrap distribution of a breakpoint
}
