% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.spec_decompose}
\alias{plot.spec_decompose}
\title{Plot a variance decomposition}
\usage{
\method{plot}{spec_decompose}(x, ...)
}
\arguments{
\item{x}{A \code{spec_decompose} object.}

\item{...}{Ignored.}
}
\value{
A ggplot object (patchwork when a null is present).
}
\description{
Plot a variance decomposition
}
