% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.spec_defensible}
\alias{plot.spec_defensible}
\title{Plot false-positive rates by level}
\usage{
\method{plot}{spec_defensible}(x, ...)
}
\arguments{
\item{x}{A \code{spec_defensible} object.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Plot false-positive rates by level
}
